refactor: organiza formadores.js

+++ src/Pages/ListaClientes.jsx
@@ -1,3 +1,4 @@
+import { useLocation } from "react-router-dom";
 
 import { useEffect, useState } from "react";
 import { supabase } from "../api/supabaseClient";
@@ -7,9 +8,21 @@ import Header from "../Componentes/Header/Header";
 
 import { createLogger } from "../lib/logger";
 const logger = createLogger("ListaClientes")
-//import EditarCliente from "../EditarCliente/editarCliente";
 
+import { formatarDataBR, formatarCEP, formatarTelefoneBR } from "../Componentes/Utilitarios/formadores";
 const ListaClientes = () => {
+
+  const location = useLocation();
+
+  const showUpdated = location.state?.updated;
+
+  useEffect(() => {
+  if (showUpdated) {
+    window.history.replaceState({}, document.title);
+  }
+}, [showUpdated]);
+
+
   const navigate = useNavigate();
   const [clientes, setClientes] = useState([]);
 
@@ -26,9 +39,13 @@ const ListaClientes = () => {
   }, []);
 
   // 🔹 Função para excluir um cliente
-  const handleExcluir = async (id) => {
-   
+  const handleExcluir = async (id, nome) => {
 
+     const confirmar = window.confirm(
+    `Tem certeza que deseja excluir o cliente "${nome}"?\n\nEssa ação não pode ser desfeita.`
+  );
+
+  if (!confirmar) return;
     const { error } = await supabase.from("clientes").delete().eq("id", id);
     if (error) {
       logger.error("Erro ao excluir cliente:", error);
@@ -54,7 +71,13 @@ const ListaClientes = () => {
 
     {/* Tabela Responsiva */}
     <div className="w-full max-w-[100%] mx-auto p-4  rounded-lg overflow-auto max-h-[500px]">
-      <div className="overflow-x-auto">
+        <div className="overflow-x-auto">
+          {showUpdated && (
+          <div className="mb-4 rounded-lg  text-green-800 font-semibold px-4 py-2  shadow">
+              ✅ Cliente atualizado com sucesso!
+        </div>
+        )}
+
         <table className="w-full border-collapse border bg-white">
           <thead>
             <tr className="border bg-gray-100 text-center text-primary font-extrabold text-sm uppercase">
@@ -75,14 +98,14 @@ const ListaClientes = () => {
             {clientes.map((cliente) => (
               <tr key={cliente.id} className="border   transition">
                 <td className="border p-2 min-w-[200px]">{cliente.nome}</td>
-                <td className="border p-2 min-w-[100px]">{cliente.data_aniversario}</td>
-                <td className="border p-2 min-w-[150px]">{cliente.telefone}</td>
+                <td className="border p-2 min-w-[100px]">{formatarDataBR(cliente.data_aniversario)}</td>
+                <td className="border p-2 min-w-[150px]">{formatarTelefoneBR(cliente.telefone)}</td>
                 <td className="border p-2">{cliente.rua}</td>
                 <td className="border p-2">{cliente.numero}</td>
                 <td className="border p-2">{cliente.complemento}</td>
                 <td className="border p-2">{cliente.bairro}</td>
                 <td className="border p-2">{cliente.cidade}</td>
-                <td className="border p-2">{cliente.cep}</td>
+                <td className="border p-2">{formatarCEP(cliente.cep)}</td>
                 <td className="border p-2 text-center">
                   <button
                     onClick={() => handleEditar(cliente.id)}
@@ -93,7 +116,7 @@ const ListaClientes = () => {
                 </td>
                 <td className="border-2 border-roxo px-3 py-2 text-center">
                   <button
-                    onClick={() => handleExcluir(cliente.id)}
+                    onClick={() => handleExcluir(cliente.id, cliente.nome)}
                     className="text-red-500 hover:text-red-700 text-xl"
                   >
                     ❌
